+++ changelog.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file
 
 ## v1.8.0
 - Fix devDependencies.
-- Asynchronous and Non-Blocking call of event listeners, even when listenrs throws Error.
+- Asynchronous and Non-Blocking call of event listeners, even when listener callback throws Error.
 - **Topic.addListener()** accepts error callback function as third argument. 
   By default the Error is sent to the console.
 - Rename **Eventset.Topic()** is deprecated and will be removed, use **Eventset.createTopic()** instead.